refactor: centralize vitest config and organize tests in __tests__ folders
- Centralize Vitest configuration to root level (vitest.config.ts)
  - Remove individual vitest.config.ts from client/ and server/
  - Update package.json scripts to use centralized commands
  - Maintain server-specific setup file (vitest.setup.ts) at root

- Reorganize test files into __tests__ folders
  - Move client/client.test.ts to client/__tests__/client.test.ts
  - Move client/index.test.ts to client/__tests__/index.test.ts
  - Move server/app.test.ts to server/__tests__/app.test.ts
  - Move server/index.test.ts to server/__tests__/index.test.ts
  - Update all import paths from './' to '../' in test files

- Update documentation to reflect new structure
  - README.md: Updated project structure diagram
  - CONTRIBUTING.md: Added note about centralized config and __tests__ folders
  - CLAUDE.md: Added centralized config information

- Maintain all existing functionality
  - All 45 tests still pass (25 client + 20 server)
  - Individual test commands work: yarn test:client, yarn test:server
  - Comprehensive mocking and test coverage preserved
