+++ package.json
@@ -7,9 +7,9 @@
     "server:start": "yarn workspace caddy-handle-scraper-server run start",
     "handshake": "node ./handshake-generator.js",
     "format": "yarn prettier -w .",
-    "test": "yarn workspaces foreach --all run test",
-    "test:client": "yarn workspace caddy-handle-scraper-client run test",
-    "test:server": "yarn workspace caddy-handle-scraper-server run test"
+    "test": "vitest run",
+    "test:client": "vitest run client",
+    "test:server": "vitest run server"
   },
   "repository": {
     "type": "git",
