+++ vitest.config.ts
@@ -4,6 +4,7 @@ export default defineConfig({
   test: {
     environment: 'node',
     globals: true,
+    include: ['**/__tests__/**/*.test.ts'],
     setupFiles: ['./vitest.setup.ts'],
   },
 });
